Persistence Testing + Touches for FYP Demo
Persistent destruction was tested in this build. In the VR Training test location, there is a building and civilian crowd which spawn/do not spawn based on the destroyed/not destroyed state they were when the level was last visited. This can be reset within the main menu for the VR area, specifically.

Music was added to touch up the VR Training Area for the FYP Demonstration.

Map screen added access to the VR Training Area.